Refactor src/barCharts/barCharts component
Split the "barCharts" component into multiple files
Make generic barChartFactory component and specific for it statementStatistics
Fix eslint errors, add eslint:fix script, add eslint env rule
