+++ .eslintrc.json
@@ -1,5 +1,8 @@
 {
   "extends": "@cockroachlabs/eslint-config",
+  "env": {
+    "node": true
+  },
   "rules": {
     "@typescript-eslint/interface-name-prefix": "off"
   }
