+++ src/barCharts/index.ts
@@ -1,1 +1,2 @@
 export * from "./barCharts";
+export * from "./breakdowns";
